+++ src/app/modules/core/models/interfaces/storage.model.ts
@@ -1,0 +1,5 @@
+export interface StorageModel {
+  setItem: Function;
+  getItem: Function;
+  removeItem: Function;
+}