+++ src/app/modules/core/services/index.ts
@@ -1,0 +1,2 @@
+export * from './config-options.service';
+export * from './local-storage.service';