+++ src/app/modules/shared/shared.module.ts
@@ -2,10 +2,11 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ObjectToArrayPipe } from './pipes/object-to-array.pipe';
 import { HighlightDirective } from './directives/highlight.directive';
+import { ConvertCaseDirective } from './directives/convert-case.directive';
 
 @NgModule({
-  declarations: [ObjectToArrayPipe, HighlightDirective],
-  exports: [ObjectToArrayPipe, HighlightDirective],
+  declarations: [ObjectToArrayPipe, HighlightDirective, ConvertCaseDirective],
+  exports: [ObjectToArrayPipe, HighlightDirective, ConvertCaseDirective],
   imports: [
     CommonModule
   ]